Built TcpConnection and OcpMessageProcessor
* TcpConnection.cxx/hxx - owns the connection, and has convenience
functions for reading particularly sized chunks from the stream, as
well as handling the correct workflow for reading OCP messages.

* OcpMessageProcessor.cxx/hxx - the TcpConnection delegates the job
of understanding the data it reads to this object. The processor is
given a callback which it calls once it is done processing the data.

* TcpServer.cxx/hxx - doesn't do a whole lot right now, but will
handle the job of arbitrating tcp connections.

* Added tests for the above, as well as necessary scaffolds.
* Added a script to debug tests.

* Added waftools, so I can use things like cppcheck from waf