Added code to read/write OcaBlob
* Created new classes for reading/writing basic types
* Created functions to read write OcaUint16
* Moved/created functions to read/write a vector<OcaUint8>
* Created functions to read/write OcaBlob
* Created relevant tests

The intention of the BasicTypeReader/Writer is that they would
modify the buffer reference passed in so that it esentially moves
along with the data as read/written. It is, therefore, imperative
that at the start of a chain of these calls, a copy of the buffer
object is passed in to the first one if the caller wishes to maintain
information about where the buffer starts.