Began adding 'New Connection Event' functionality into OcaNetwork.
The point of this functionality is to allow the TcpServer to
notify the OcaNetwork when a new socket connection (TcpConnection)
is accepted. The OcaNetwork the creates the OcpSession object, and
will also (in due course) regsiter it with the (to be created)
OcaApplication class, which will be the main interface to the host
device.